+++ tjbench.c
@@ -90,7 +90,7 @@ int tjErrorLine = -1, tjErrorCode = -1;
 
 int stopOnWarning = 0, bottomUp = 0, noRealloc = 1, fastUpsample = 0,
   fastDCT = 0, optimize = 0, progressive = 0, limitScans = 0, maxMemory = 0,
-  arithmetic = 0, lossless = 0, restartIntervalBlocks = 0,
+  maxPixels = 0, arithmetic = 0, lossless = 0, restartIntervalBlocks = 0,
   restartIntervalRows = 0;
 int precision = 8, sampleSize, compOnly = 0, decompOnly = 0, doYUV = 0,
   quiet = 0, doTile = 0, pf = TJPF_BGR, yuvAlign = 1, doWrite = 1;
@@ -203,6 +203,8 @@ static int decomp(unsigned char **jpegBufs, size_t *jpegSizes, void *dstBuf,
   if (tj3Set(handle, TJPARAM_SCANLIMIT, limitScans ? 500 : 0) == -1)
     THROW_TJ();
   if (tj3Set(handle, TJPARAM_MAXMEMORY, maxMemory) == -1)
+    THROW_TJ();
+  if (tj3Set(handle, TJPARAM_MAXPIXELS, maxPixels) == -1)
     THROW_TJ();
 
   if (IS_CROPPED(cr)) {
@@ -659,6 +661,8 @@ static int decompTest(char *fileName)
     THROW_TJ();
   if (tj3Set(handle, TJPARAM_MAXMEMORY, maxMemory) == -1)
     THROW_TJ();
+  if (tj3Set(handle, TJPARAM_MAXPIXELS, maxPixels) == -1)
+    THROW_TJ();
 
   if (tj3DecompressHeader(handle, srcBuf, srcSize) == -1)
     THROW_TJ();
@@ -910,6 +914,7 @@ static void usage(char *progName)
   printf("     progressive JPEG compression and decompression, optimized baseline entropy\n");
   printf("     coding, lossless JPEG compression, and lossless transformation\n");
   printf("     [default = no limit]\n");
+  printf("-maxpixels = Input image size limit (in pixels) [default = no limit]\n");
   printf("-nowrite = Do not write reference or output images (improves consistency of\n");
   printf("     benchmark results)\n");
   printf("-rgb, -bgr, -rgbx, -bgrx, -xbgr, -xrgb =\n");
@@ -1161,6 +1166,11 @@ int main(int argc, char *argv[])
 
         if (tempi < 0) usage(argv[0]);
         maxMemory = tempi;
+      } else if (!strcasecmp(argv[i], "-maxpixels") && i < argc - 1) {
+        int tempi = atoi(argv[++i]);
+
+        if (tempi < 0) usage(argv[0]);
+        maxPixels = tempi;
       } else if (!strcasecmp(argv[i], "-restart") && i < argc - 1) {
         int tempi = -1, nscan;  char tempc = 0;
 
@@ -1227,6 +1237,8 @@ int main(int argc, char *argv[])
     if (tj3Set(handle, TJPARAM_STOPONWARNING, stopOnWarning) == -1)
       THROW_TJ();
     if (tj3Set(handle, TJPARAM_BOTTOMUP, bottomUp) == -1)
+      THROW_TJ();
+    if (tj3Set(handle, TJPARAM_MAXPIXELS, maxPixels) == -1)
       THROW_TJ();
 
     if (precision == 8) {